requirements.txt
flask
rethinkdb
python-socketio
eventlet